[UPDATE] training script for HHM

+++ scripts/prepare_hhm.sh
@@ -1,18 +1,20 @@
 cd /mnt/localssd
-aws s3 cp --recursive s3://chuongh-vm2m/HHM HHM
-cd HHM
+if [ ! -d "HHM" ]; then
+    aws s3 cp --recursive s3://chuongh-vm2m/HHM HHM
+    cd HHM
+    echo "Unzipping alphas and HHM2K..."
+    unzip -q alphas.zip
+    unzip -q HHM2K.zip
+    rm alphas.zip
+    rm HHM2K.zip
 
-unzip -q alphas.zip
-unzip -q HHM2K.zip
-rm alphas.zip
-rm HHM2K.zip
-
-mv images train
-mv HHM2K val
-mv alphas train/alphas
-
-cd train
-cat images.zip.* > images.zip
-unzip -q images.zip
-rm images.zip.*
-rm images.zip
+    mv images train
+    mv HHM2K val
+    mv alphas train/alphas
+    echo "Unzipping images..."
+    cd train
+    cat images.zip.* > images.zip
+    unzip -q images.zip
+    rm images.zip.*
+    rm images.zip
+fi
